requirements.txt
  
Flask==2.3.3
qrcode==7.4.2
Pillow==10.0.1
opencv-python-headless==4.8.1.78
numpy==1.24.3
gunicorn==21.2.0
Werkzeug==2.3.7